feat event : Separated event banner edit from event edit for a better UX solved map marker issue
