+++ templates/event/editBanner.html.twig
@@ -1,0 +1,10 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Modifier un événement{% endblock %}
+
+{% block body %}
+      {{ include('event/_form.html.twig', {'button_label': 'Mettre à jour'}) }}
+     <div class = "button-container">
+      {{ include('event/_delete_form.html.twig') }}
+     </div>
+{% endblock %}
